+++ poetry/installation/pip_installer.py
@@ -108,7 +108,14 @@ class PipInstaller(BaseInstaller):
     def requirement(self, package, formatted=False):
         if formatted and not package.source_type:
             req = "{}=={}".format(package.name, package.version)
-            for h in package.hashes:
+
+            # metadata.hashes is still supported as a fallback method
+            # if files has not yet been created
+            hashes = package.hashes
+            if package.files:
+                hashes = (f["hash"] for f in package.files)
+
+            for h in hashes:
                 hash_type = "sha256"
                 if ":" in h:
                     hash_type, h = h.split(":")